populate: correct bug with casting float as int

+++ MEBS/src/populate.C
@@ -47,7 +47,7 @@ float back_to_back(float photon_phi, float jet_phi, float threshold) {
 
 void fill_axes(pjtree* pjt, multival* mpthf, multival* mpthfjpt, multival* mpthfeta, multival* mpthfdphi, 
                interval *ijpt, interval *ieta, interval *idphi, int64_t pt_x, int64_t hf_x,
-               float weight, int64_t photon_phi, bool heavyion, 
+               float weight, float photon_phi, bool heavyion, 
                float jet_eta_abs, float jet_dr_max, float jet_pt_min, float jet_pt_max,
                float dphi_min_numerator, float dphi_min_denominator, memory<TH1F>* nevt, 
                memory<TH1F>* jpt_pjet_f_dr, memory<TH1F>* jpt_pjet_f_jpt,
